+++ internal/vehicles/domain/commands/get_vehicle_by_id_command.go
@@ -1,29 +1,29 @@
 package commands
 
 import (
-  "errors"
+	"errors"
 
-  de "github.com/kalilventura/vehicle-management/internal/shared/domain/errors"
-  "github.com/kalilventura/vehicle-management/internal/vehicles/domain/repositories"
+	domainerror "github.com/kalilventura/vehicle-management/internal/shared/domain/errors"
+	"github.com/kalilventura/vehicle-management/internal/vehicles/domain/repositories"
 )
 
 type GetVehicleByIDCommand struct {
-  repository repositories.VehiclesRepository
+	repository repositories.VehiclesRepository
 }
 
 func NewGetVehicleByIDCommand(repository repositories.VehiclesRepository) *GetVehicleByIDCommand {
-  return &GetVehicleByIDCommand{repository}
+	return &GetVehicleByIDCommand{repository}
 }
 
 func (cmd *GetVehicleByIDCommand) Execute(ID string, listeners GetVehicleByIDListeners) {
-  target, err := cmd.repository.GetByID(ID)
-  switch {
-  case errors.Is(err, de.ErrRecordNotFound):
-    listeners.OnNotFound()
-    return
-  case err != nil:
-    listeners.OnInternalServerError(err)
-    return
-  }
-  listeners.OnSuccess(target)
+	target, err := cmd.repository.GetByID(ID)
+	switch {
+	case errors.Is(err, domainerror.ErrRecordNotFound):
+		listeners.OnNotFound()
+		return
+	case err != nil:
+		listeners.OnInternalServerError(err)
+		return
+	}
+	listeners.OnSuccess(target)
 }