+++ internal/vehicles/domain/entities/dtos/list_vehicles_input.go
@@ -1,0 +1,14 @@
+package dtos
+
+import (
+	"github.com/kalilventura/vehicle-management/internal/shared/domain/entities"
+)
+
+type ListVehiclesInput struct {
+	Status     *Status
+	MinPrice   *Price
+	MaxPrice   *Price
+	Pagination entities.Pagination
+	SortBy     string
+	SortOrder  string
+}