+++ internal/vehicles/domain/entities/dtos/price.go
@@ -5,8 +5,8 @@ import "errors"
 type Price float64
 
 func NewPrice(value float64) (Price, error) {
-	if value < 0 {
-		return 0, errors.New("price cannot be negative")
+	if value <= 0 {
+		return 0, errors.New("price cannot be negative and needs to be greater than 0")
 	}
 	return Price(value), nil
 }