+++ internal/vehicles/infrastructure/controllers/container.go
@@ -8,4 +8,5 @@ import (
 var Container = wire.NewSet(
 	NewSaveVehiclesController,
 	NewGetVehicleByIdController,
+	NewListVehiclesController,
 )